% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{existsProjectOption,character,ANY-method}
\alias{existsProjectOption,character,ANY-method}
\alias{existsProjectOption-char-any-method}
\title{Check Existence of Project Option (char-any)}
\usage{
\S4method{existsProjectOption}{character,ANY}(id,
  where = tryCatch(devtools::as.package(".")$package, error = function(cond) {
      stop("Invalid default value for `where`") }), sub_id = character(),
  default = NULL, strict = c(0, 1, 2), ...)
}
\arguments{
\item{id}{\code{\link{character}}.}

\item{where}{\code{\link{ANY}}.}
}
\value{
See method
   \code{\link[optionr]{existsProjectOption-char-env-method}}
}
\description{
See generic: \code{\link[optionr]{existsProjectOption}}
}
\examples{
\dontrun{

##------------------------------------------------------------------------------
## Basics //
##------------------------------------------------------------------------------

setProjectOption(id = "test", value = TRUE)
existsProjectOption(id = "test")

setProjectOption(id = "a/b/c", value = 10)
existsProjectOption(id = "a")
existsProjectOption(id = "a/b")
existsProjectOption(id = "a/b/c")
existsProjectOption(id = "a/b/c/d")

existsProjectOption(id = "c")
existsProjectOption(id = "c/d/e")
  
##------------------------------------------------------------------------------
## Strictness levels //
##------------------------------------------------------------------------------

## Empty ID //
existsProjectOption(id = character())
try(existsProjectOption(id = character(), strict = 1))
try(existsProjectOption(id = character(), strict = 2))

## Not-existing //  
existsProjectOption(id = "c/d/e")
try(existsProjectOption(id = "c/d/e", strict = 1))
try(existsProjectOption(id = "c/d/e", strict = 2))

##------------------------------------------------------------------------------
## Explicit `where` //
##------------------------------------------------------------------------------

where <- new.env()
setProjectOption(id = "a/b/c", value = 10, where = where)
existsProjectOption(id = "a/b/c", where = where)
existsProjectOption(id = "a/b/c/d", where = where)
existsProjectOption(id = "c/d/e", where = where)

##------------------------------------------------------------------------------
## Sub ID //
##------------------------------------------------------------------------------

where <- new.env()
setProjectOption(id = "a/b/c", value = 10, where = where, sub_id = "sub")
existsProjectOption(id = "a/b/c", where = where, sub_id = "sub")
existsProjectOption(id = "a/b/c/d", where = where, sub_id = "sub")
existsProjectOption(id = "c/d/e", where = where, sub_id = "sub")

}
}
\author{
Janko Thyson \email{janko.thyson@gmail.com}
}
\references{
\url{http://github.com/Rappster/optionr}
}
\seealso{
\code{
   \link[optionr]{existsProjectOption}
}
}

