% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{existsProjectOption,character,missing-method}
\alias{existsProjectOption,character,missing-method}
\alias{existsProjectOption-char-miss-method}
\title{Check Existence of Project Option (char-miss)}
\usage{
\S4method{existsProjectOption}{character,missing}(id,
  where = tryCatch(devtools::as.package(".")$package, error = function(cond) {
      stop("Invalid default value for `where`") }), sub_id = character(),
  default = NULL, strict = c(0, 1, 2), ...)
}
\arguments{
\item{id}{\code{\link{character}}.}

\item{where}{\code{\link{missing}}.}

\item{sub_id}{\code{\link{character}}.
Optional ID for a sub layer. Useful for a hub-like option container
structure.}

\item{default}{\code{\link{ANY}}.
Value to be returned if option does not exist.
See \code{\link[base]{getOption}} and \code{\link[nestr]{getNested}}.}

\item{strict}{\code{\link{logical}}.
 Controls what happens when \code{id} points to a non-existing option object:
\itemize{
     \item{0: }{ignore and return \code{FALSE} to signal that the
             assignment process was not successful or \code{fail_value} depending
             on the value of \code{return_status}}
         \item{1: }{ignore and with warning and return \code{FALSE}}
         \item{2: }{ignore and with error}
   }}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
See method
   \code{\link[optionr]{existsProjectOption-char-char-method}}
}
\description{
See generic: \code{\link[optionr]{existsProjectOption}}
}
\examples{
\dontrun{

##------------------------------------------------------------------------------
## Basics //
##------------------------------------------------------------------------------

setProjectOption(id = "test", value = TRUE)
existsProjectOption(id = "test")

setProjectOption(id = "a/b/c", value = 10)
existsProjectOption(id = "a")
existsProjectOption(id = "a/b")
existsProjectOption(id = "a/b/c")
existsProjectOption(id = "a/b/c/d")

existsProjectOption(id = "c")
existsProjectOption(id = "c/d/e")
  
##------------------------------------------------------------------------------
## Strictness levels //
##------------------------------------------------------------------------------

## Empty ID //
existsProjectOption(id = character())
try(existsProjectOption(id = character(), strict = 1))
try(existsProjectOption(id = character(), strict = 2))

## Not-existing //  
existsProjectOption(id = "c/d/e")
try(existsProjectOption(id = "c/d/e", strict = 1))
try(existsProjectOption(id = "c/d/e", strict = 2))

##------------------------------------------------------------------------------
## Explicit `where` //
##------------------------------------------------------------------------------

where <- new.env()
setProjectOption(id = "a/b/c", value = 10, where = where)
existsProjectOption(id = "a/b/c", where = where)
existsProjectOption(id = "a/b/c/d", where = where)
existsProjectOption(id = "c/d/e", where = where)

##------------------------------------------------------------------------------
## Sub ID //
##------------------------------------------------------------------------------

where <- new.env()
setProjectOption(id = "a/b/c", value = 10, where = where, sub_id = "sub")
existsProjectOption(id = "a/b/c", where = where, sub_id = "sub")
existsProjectOption(id = "a/b/c/d", where = where, sub_id = "sub")
existsProjectOption(id = "c/d/e", where = where, sub_id = "sub")

}
}
\author{
Janko Thyson \email{janko.thyson@gmail.com}
}
\references{
\url{http://github.com/Rappster/optionr}
}
\seealso{
\code{
   \link[optionr]{existsProjectOption}
}
}

