% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getAnywhereOption}
\alias{getAnywhereOption}
\title{Get Anywhere Option (generic)}
\usage{
getAnywhereOption(id, where = tryCatch(devtools::as.package(".")$package,
  error = function(cond) {     stop("Invalid default value for `where`") }),
  default = NULL, strict = FALSE, ...)
}
\arguments{
\item{id}{\strong{Signature argument}.
Object containing path-like name/ID information.}

\item{where}{\strong{Signature argument}.
Object containing information about the location of the option container
that is to be used. Typically, this either corresponds to the name/ID
of a package/package project or an instance of a custom class for which
suitable methods in the context of managing options are defined
(see other methods of this package that have signature arguments
\code{id} or \code{where}).}

\item{default}{\code{\link{ANY}}.
Value to be returned if option does not exist.
See \code{\link[base]{getOption}}.}

\item{strict}{\code{\link{logical}}.
\code{TRUE}: \code{id} pointing to a non-existing option triggers
error; \code{FALSE}: \code{id} pointing to a non-existing option leads
to return value \code{NULL}.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\description{
Retrieves option from anywhere inside an option container
based on a path-like identifier (e.g. \code{"container/subcontainer/option_name"}.
}
\section{Path-like identifiers}{


Values for \code{id} are expected to be of structure \code{a/b/c/.../z},
i.e. path-like identifiers using a slash as separator.
The identifier is transformed to \code{a$b$c$...$z} and then in turn to a
valid \emph{get} or \code{assign} expression
(i.e. something similar to \code{getOptionContainer(...)$a$b$c$...$z} and
\code{getOptionContainer(...)$a$b$c$...$z <- value}).
Of course, "atomic" paths (e.g. only \code{a}) are also valid.
}
\examples{
\dontrun{

## Also see examples at `?setAnywhereOption`
  
##------------------------------------------------------------------------------
## Basics //
##------------------------------------------------------------------------------

container <- initializeOptionContainer(overwrite = TRUE)
setAnywhereOption(id = "test", value = TRUE)
getAnywhereOption(id = "test")

setAnywhereOption(id = "a/b/c", value = 10, gap = TRUE)
getAnywhereOption(id = "a")
getAnywhereOption(id = "a/b")
getAnywhereOption(id = "a/b/c")
  
##------------------------------------------------------------------------------
## Different `where` //
##------------------------------------------------------------------------------

where <- "test"
container <- initializeOptionContainer(id = where, overwrite = TRUE)
setAnywhereOption(id = "a/b/c", value = 10, where = where, gap = TRUE)
getAnywhereOption(id = "a/b/c", where = where)

where <- structure(list(id = "test"), class = "OptionContext.Test")
container <- initializeOptionContainer(id = where, overwrite = TRUE)
setAnywhereOption(id = "a/b/c", value = 10, where = where, gap = TRUE)
getAnywhereOption(id = "a/b/c", where = where)
  
}
}
\author{
Janko Thyson \email{janko.thyson@gmail.com}
}
\references{
\url{http://github.com/Rappster/optionr}
}
\seealso{
\code{
  	\link[optionr]{getAnywhereOption-char-char-method},
    \link[optionr]{setAnywhereOption},
    \link[optionr]{rmAnywhereOption}
}
}

