% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{getFreeOption,character,character-method}
\alias{getFreeOption,character,character-method}
\alias{getFreeOption-char-miss-method}
\title{Get Generic Option (char-char)}
\usage{
\S4method{getFreeOption}{character,character}(id,
  where = tryCatch(devtools::as.package(".")$package, error = function(cond) {
      stop("Invalid default value for `where`") }), default = NULL,
  strict = FALSE, ...)
}
\arguments{
\item{id}{\code{\link{character}}.}

\item{where}{\code{\link{character}}.}

\item{default}{\code{\link{ANY}}.
Value to be returned if option does not exist.
See \code{\link[base]{getOption}}.}

\item{strict}{\code{\link{logical}}.
\code{TRUE}: \code{id} pointing to a non-existing option triggers
error; \code{FALSE}: \code{id} pointing to a non-existing option leads
to return value \code{NULL}.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
\code{\link{ANY}}. Option value or for non-existing option
   (i.e. wrong \code{id}): \code{NULL} if \code{strict = FALSE} and an error
   if \code{strict = TRUE}.
}
\description{
See generic: \code{\link[optionr]{getFreeOption}}
}
\examples{
\dontrun{

## Also see examples at `?setFreeOption`
  
##------------------------------------------------------------------------------
## Basics //
##------------------------------------------------------------------------------

container <- initializeOptionContainer(overwrite = TRUE)
setFreeOption(id = "test", value = TRUE)
getFreeOption(id = "test")

setFreeOption(id = "a/b/c", value = 10, gap = TRUE)
getFreeOption(id = "a")
getFreeOption(id = "a/b")
getFreeOption(id = "a/b/c")
  
##------------------------------------------------------------------------------
## Different `where` //
##------------------------------------------------------------------------------

where <- "test"
container <- initializeOptionContainer(id = where, overwrite = TRUE)
setFreeOption(id = "a/b/c", value = 10, where = where, gap = TRUE)
getFreeOption(id = "a/b/c", where = where)

where <- structure(list(id = "test"), class = "OptionContext.Test")
container <- initializeOptionContainer(id = where, overwrite = TRUE)
setFreeOption(id = "a/b/c", value = 10, where = where, gap = TRUE)
getFreeOption(id = "a/b/c", where = where)
  
}
}
\author{
Who to complain to \email{yourfault@somewhere.net}
}
\references{
\url{http://github.com/whotocomplainto/test.package}
}
\seealso{
\code{
   \link[optionr]{getFreeOption}
}
}

