% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{getFreeOption,character-method}
\alias{getFreeOption,character-method}
\title{Get Generic Option (char)}
\usage{
\S4method{getFreeOption}{character}(id, strict = FALSE, ...)
}
\arguments{
\item{id}{\code{\link{character}}.}

\item{strict}{\code{\link{logical}}.
\code{TRUE}: \code{id} pointing to a non-existing option triggers
error; \code{FALSE}: \code{id} pointing to a non-existing option leads
to return value \code{NULL}.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
\code{\link{ANY}}. Option value or for non-existing option
   (i.e. wrong \code{id}): \code{NULL} if \code{strict = FALSE} and an error
   if \code{strict = TRUE}.
}
\description{
See generic: \code{\link[optionr]{getFreeOption}}
}
\examples{
\dontrun{

## TODO: add example

}
}
\author{
Who to complain to \email{yourfault@somewhere.net}
}
\references{
\url{http://github.com/whotocomplainto/test.package}
}
\seealso{
\code{
   \link[optionr]{getFreeOption}
}
}

