% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{getProjectOption,character,ANY-method}
\alias{getProjectOption,character,ANY-method}
\alias{getProjectOption-char-any-method}
\title{Get Option (char-any)}
\usage{
\S4method{getProjectOption}{character,ANY}(id,
  where = tryCatch(devtools::as.package(".")$package, error = function(cond) {
      stop("Invalid default value for `where`") }), default = NULL,
  strict = FALSE, ...)
}
\arguments{
\item{id}{\code{\link{character}}.}

\item{where}{\code{\link{ANY}}.}

\item{default}{\code{\link{ANY}}.
Value to be returned if option does not exist.
See \code{\link[base]{getProjectOption}}.}

\item{strict}{\code{\link{logical}}.
\code{TRUE}: \code{id} pointing to a non-existing option triggers
error; \code{FALSE}: \code{id} pointing to a non-existing option leads
to return value \code{NULL}.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
See method
   \code{\link[optionr]{getProjectOption-char-char-method}}
}
\description{
See generic: \code{\link[optionr]{getProjectOption}}
}
\examples{
\dontrun{

## TODO: add example

}
}
\author{
Who to complain to \email{yourfault@somewhere.net}
}
\references{
\url{http://github.com/whotocomplainto/test.package}
}
\seealso{
\code{
   \link[optionr]{getProjectOption}
}
}

