% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{getProjectOption,character,character-method}
\alias{getProjectOption,character,character-method}
\alias{getProjectOption-char-miss-method}
\title{Get Project Option (char-char)}
\usage{
\S4method{getProjectOption}{character,character}(id,
  where = tryCatch(devtools::as.package(".")$package, error = function(cond) {
      stop("Invalid default value for `where`") }), default = NULL,
  strict = c(0, 1, 2), ...)
}
\arguments{
\item{id}{\code{\link{character}}.}

\item{where}{\code{\link{character}}.}

\item{default}{\code{\link{ANY}}.
Value to be returned if option does not exist.
See \code{\link[base]{getOption}} and \code{\link[nestr]{getNested}}.}

\item{strict}{\code{\link{logical}}.
 Controls what happens when \code{id} points to a non-existing option object:
\itemize{
     \item{0: }{ignore and return \code{FALSE} to signal that the
             assignment process was not successful or \code{fail_value} depending
             on the value of \code{return_status}}
         \item{1: }{ignore and with warning and return \code{FALSE}}
         \item{2: }{ignore and with error}
   }}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
\code{\link{ANY}}. Component value or for invalid argument input
		and non-existing component the value of \code{default} unless
		\code{strict == 1} in which case a warning is issued or
		\code{strict == 2} in which case an error is thrown.
}
\description{
See generic: \code{\link[optionr]{getProjectOption}}
}
\examples{
\dontrun{

## TODO: add example

}
}
\author{
Janko Thyson \email{janko.thyson@gmail.com}
}
\references{
\url{http://github.com/Rappster/optionr}
}
\seealso{
\code{
   \link[optionr]{getProjectOption}
}
}

