% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getRegistryValue}
\alias{getRegistryValue}
\title{Get Registry Value (generic)}
\usage{
getRegistryValue(id, where = tryCatch(devtools::as.package(".")$package, error
  = function(cond) {     stop("Invalid default value for `where`") }),
  default = NULL, strict = FALSE, ...)
}
\arguments{
\item{id}{\strong{Signature argument}.
Object containing path-like name/ID information.}

\item{where}{\strong{Signature argument}.
Object containing information about the location of the option container
that is to be used. Typically, this either corresponds to the name/ID
of a package/package project or an instance of a custom class for which
suitable methods in the context of managing options are defined
(see other methods of this package that have signature arguments
\code{id} or \code{where}).}

\item{default}{\code{\link{ANY}}.
Value to be returned if option does not exist.
See \code{\link[base]{getRegistryValue}}.}

\item{strict}{\code{\link{logical}}.
\code{TRUE}: \code{id} pointing to a non-existing option triggers
error; \code{FALSE}: \code{id} pointing to a non-existing option leads
to return value \code{NULL}.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\description{
Retrieves registry value from the respective environment for registry information
inside an option container (see \code{\link[optionr]{initializeOptionContainer}}).
}
\section{Path-like identifiers}{


Values for \code{id} are expected to be of structure \code{a/b/c/.../z},
i.e. path-like identifiers using a slash as separator.
The identifier is transformed to \code{a$b$c$...$z} and then in turn to a
valid \emph{get} or \code{assign} expression
(i.e. something similar to \code{getOptionContainer(...)$a$b$c$...$z} and
\code{getOptionContainer(...)$a$b$c$...$z <- value}).
Of course, "atomic" paths (e.g. only \code{a}) are also valid.
}
\examples{
\dontrun{

## TODO: add example

}
}
\author{
Janko Thyson \email{janko.thyson@gmail.com}
}
\references{
\url{http://github.com/Rappster/optionr}
}
\seealso{
\code{
  	\link[optionr]{getRegistryValue-char-char-method},
   \link[optionr]{setRegistryValue},
   \link[optionr]{rmRegistryValue}
}
}

