% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{initializeMeta,missing,environment-method}
\alias{initializeMeta,missing,environment-method}
\alias{initializeMeta-missing-environment-method}
\title{Initialize Meta Information (miss-envir)}
\usage{
\S4method{initializeMeta}{missing,environment}(main = ".meta", where, ...)
}
\arguments{
\item{main}{\code{\link{missing}}.}

\item{where}{\code{\link{environment}}.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
See method
   \code{\link[reactr]{initializeMeta-NULL-method}}.
}
\description{
See generic: \code{\link[reactr]{initializeMeta}}
}
\examples{
\dontrun{

## TODO: add example

}
}
\author{
Who to complain to \email{yourfault@somewhere.net}
}
\references{
\url{http://github.com/whotocomplainto/test.package}
}
\seealso{
\code{
   \link[reactr]{initializeMeta-NULL-method},
    \link[reactr]{getRegistry}
}
}

