% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{initializeMeta,missing,missing-method}
\alias{initializeMeta,missing,missing-method}
\alias{initializeMeta-miss-miss-method}
\title{Initialize Meta Information (miss-miss)}
\usage{
\S4method{initializeMeta}{missing,missing}(id = ".meta", where, ...)
}
\arguments{
\item{id}{\code{\link{missing}}.}

\item{where}{\code{\link{missing}}.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
See method
   \code{\link[optionr]{initializeMeta-char-env-method}}.
}
\description{
See generic: \code{\link[optionr]{initializeMeta}}
}
\examples{
\dontrun{

## Default `id` //  
container <- ensureOptionContainer(overwrite = TRUE)  
initializeMeta(where = container)
exists(".meta", container)

## Different `id` //
container <- ensureOptionContainer(overwrite = TRUE)  
initializeMeta(id = "META", where = container)
exists("META", container)

}
}
\author{
Janko Thyson \email{janko.thyson@gmail.com}
}
\references{
\url{http://github.com/Rappster/optionr}
}
\seealso{
\code{
   \link[optionr]{initializeMeta-char-env-method},
   \link[optionr]{setMetaValue},
   \link[optionr]{getMetaValue},
   \link[optionr]{rmMetaValue}
}
}

