% Generated by roxygen2 (4.0.2): do not edit by hand
\name{initializeMeta}
\alias{initializeMeta}
\title{Initialize Meta Information (generic)}
\usage{
initializeMeta(id = ".meta", where, ...)
}
\arguments{
\item{id}{\strong{Signature argument}.
Object containing ID information.}

\item{where}{\strong{Signature argument}.
Object containing location information.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\description{
Convenience function to initialize a meta information environment inside an
option container.
}
\examples{
\dontrun{

## Default `id` //  
container <- ensureOptionContainer(overwrite = TRUE)  
initializeMeta(where = container)
exists(".meta", container)

## Different `id` //
container <- ensureOptionContainer(overwrite = TRUE)  
initializeMeta(id = "META", where = container)
exists("META", container)

}
}
\author{
Who to complain to \email{yourfault@somewhere.net}
}
\references{
\url{http://github.com/whotocomplainto/test.package}
}
\seealso{
\code{
  	\link[reactr]{initializeMeta-NULL-method},
    \link[reactr]{getRegistry}
}
}

