% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{initializeOptions,character,environment-method}
\alias{initializeOptions,character,environment-method}
\alias{initializeOptions-character-environment-method}
\title{Initialize Options (char-envir)}
\usage{
\S4method{initializeOptions}{character,environment}(id = "options", where,
  ...)
}
\arguments{
\item{id}{\code{\link{character}}.}

\item{where}{\code{\link{environment}}.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
\code{\link{environment}}. Environment containing the options.
}
\description{
See generic: \code{\link[reactr]{initializeOptions}}
}
\examples{
\dontrun{

## Default `id` //  
container <- ensureOptionContainer(overwrite = TRUE)  
initializeOptions(where = container)
exists("options", container)

## Different `id` //
container <- ensureOptionContainer(overwrite = TRUE)  
initializeOptions(id = "OPTIONS", where = container)
exists("OPTIONS", container)

}
}
\author{
Who to complain to \email{yourfault@somewhere.net}
}
\references{
\url{http://github.com/whotocomplainto/test.package}
}
\seealso{
\code{
   \link[reactr]{initializeOptions},
    \link[reactr]{getRegistry}
}
}

