% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{initializeProjectOptions,ANY,environment-method}
\alias{initializeProjectOptions,ANY,environment-method}
\alias{initializeProjectOptions-any-env-method}
\title{Initialize Options (any-envir)}
\usage{
\S4method{initializeProjectOptions}{ANY,environment}(id = "options", where,
  ...)
}
\arguments{
\item{id}{\code{\link{ANY}}.}

\item{where}{\code{\link{environment}}.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
\code{\link{environment}}. Environment containing the options.
}
\description{
See generic: \code{\link[optionr]{initializeProjectOptions}}
}
\examples{
\dontrun{

## Default `id` //  
container <- ensureOptionContainer(overwrite = TRUE)  
initializeProjectOptions(where = container)
exists("options", container)

## Different `id` //
container <- ensureOptionContainer(overwrite = TRUE)  
initializeProjectOptions(id = "OPTIONS", where = container)
exists("OPTIONS", container)

}
}
\author{
Janko Thyson \email{janko.thyson@gmail.com}
}
\references{
\url{http://github.com/Rappster/optionr}
}
\seealso{
\code{
   \link[optionr]{initializeProjectOptions}
}
}

