% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{initializeProjectOptions,missing,environment-method}
\alias{initializeProjectOptions,missing,environment-method}
\alias{initializeProjectOptions-miss-env-method}
\title{Initialize Options (miss-envir)}
\usage{
\S4method{initializeProjectOptions}{missing,environment}(id = "options",
  where, ...)
}
\arguments{
\item{id}{\code{\link{missing}}.}

\item{where}{\code{\link{environment}}.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
See method
   \code{\link[optionr]{initializeProjectOptions-char-env-method}}.
}
\description{
See generic: \code{\link[optionr]{initializeProjectOptions}}
}
\examples{
\dontrun{

## Default `id` //  
container <- ensureOptionContainer(overwrite = TRUE)  
initializeProjectOptions(where = container)
exists("options", container)

## Different `id` //
container <- ensureOptionContainer(overwrite = TRUE)  
initializeProjectOptions(id = "OPTIONS", where = container)
exists("OPTIONS", container)

}
}
\author{
Janko Thyson \email{janko.thyson@gmail.com}
}
\references{
\url{http://github.com/Rappster/optionr}
}
\seealso{
\code{
   \link[optionr]{initializeProjectOptions}
}
}

