% Generated by roxygen2 (4.0.2): do not edit by hand
\name{initializeProjectOptions}
\alias{initializeProjectOptions}
\title{Initialize Options (generic)}
\usage{
initializeProjectOptions(id = "options", where, ...)
}
\arguments{
\item{id}{\strong{Signature argument}.
Object containing ID information.}

\item{where}{\strong{Signature argument}.
Object containing location information.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\description{
Convenience function to initialize a package options environment below an
option container.
}
\examples{
\dontrun{

## Default `id` //  
container <- ensureOptionContainer(overwrite = TRUE)  
initializeProjectOptions(where = container)
exists("options", container)

## Different `id` //
container <- ensureOptionContainer(overwrite = TRUE)  
initializeProjectOptions(id = "OPTIONS", where = container)
exists("OPTIONS", container)

}
}
\author{
Janko Thyson \email{janko.thyson@gmail.com}
}
\references{
\url{http://github.com/Rappster/optionr}
}
\seealso{
\code{
  	\link[optionr]{initializeProjectOptions-char-env-method},
   \link[optionr]{setProjectOptions},
   \link[optionr]{getProjectOptions},
   \link[optionr]{rmProjectOptions}
}
}

