% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{initializeRegistry,ANY,environment-method}
\alias{initializeRegistry,ANY,environment-method}
\alias{initializeRegistry-any-environment-method}
\title{Initialize Registry (any-envir)}
\usage{
\S4method{initializeRegistry}{ANY,environment}(main = ".registry", where, ...)
}
\arguments{
\item{main}{\code{\link{ANY}}.}

\item{where}{\code{\link{environment}}.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
\code{\link{environment}}. Initialized registry.
}
\description{
See generic: \code{\link[reactr]{initializeRegistry}}
}
\examples{
\dontrun{

## TODO: add example

}
}
\author{
Who to complain to \email{yourfault@somewhere.net}
}
\references{
\url{http://github.com/whotocomplainto/test.package}
}
\seealso{
\code{
   \link[reactr]{initializeRegistry},
    \link[reactr]{getRegistry}
}
}

