% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{initializeRegistry,character,environment-method}
\alias{initializeRegistry,character,environment-method}
\alias{initializeRegistry-char-env-method}
\title{Initialize Registry (char-envir)}
\usage{
\S4method{initializeRegistry}{character,environment}(id = ".registry", where,
  ...)
}
\arguments{
\item{id}{\code{\link{character}}.}

\item{where}{\code{\link{environment}}.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
\code{\link{environment}}. Initialized registry.
}
\description{
See generic: \code{\link[optionr]{initializeRegistry}}
}
\examples{
\dontrun{

## Default `id` //  
container <- ensureOptionContainer(overwrite = TRUE)  
initializeRegistry(where = container)
exists(".registry", container)

## Different `id` //
container <- ensureOptionContainer(overwrite = TRUE)  
initializeRegistry(id = "REGISTRY", where = container)
exists("REGISTRY", container)

}
}
\author{
Janko Thyson \email{janko.thyson@gmail.com}
}
\references{
\url{http://github.com/Rappster/optionr}
}
\seealso{
\code{
   \link[optionr]{initializeRegistry},
    \link[optionr]{getRegistry}
}
}

