% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{initializeRuntimeOptions,missing-method}
\alias{initializeRuntimeOptions,missing-method}
\title{Initialize Runtime Options (miss)}
\usage{
\S4method{initializeRuntimeOptions}{missing}(ns, ...)
}
\arguments{
\item{ns}{\code{\link{missing}}. Default namespace.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
\code{\link{environment}}. The options container as stored in
   \code{options(".rapp")}.
}
\description{
See generic: \code{\link[rapp]{initializeRuntimeOptions}}
}
\examples{
\dontrun{

## TODO: add example

}
}
\author{
Who to complain to \email{yourfault@somewhere.net}
}
\references{
\url{http://github.com/whotocomplainto/test.package}
}
\seealso{
\code{
   \link[rapp]{initializeRuntimeOptions}
}
}

