% Generated by roxygen2 (4.0.2): do not edit by hand
\name{initializeRuntimeOptions}
\alias{initializeRuntimeOptions}
\title{Initialize Runtime Options (generic)}
\usage{
initializeRuntimeOptions(ns, ...)
}
\arguments{
\item{ns}{\strong{Signature argument}.
Object containing namespace information.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\description{
Ensures initial options as expected by the \code{rapp} framework and its
associated packages. The options are stored as \code{".rapp"} in
\code{\link[base]{options}}.
}
\examples{
\dontrun{

## TODO: add example

}
}
\author{
Who to complain to \email{yourfault@somewhere.net}
}
\references{
\url{http://github.com/whotocomplainto/test.package}
}
\seealso{
\code{
  	\link[rapp]{initializeRuntimeOptions-missing-method}
}
}

