% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{optionr}
\alias{optionr}
\alias{optionr-package}
\title{Package Title}
\description{
Package description
}
\details{
The core functions/methods of this package:
 \itemize{
   \item{\code{\link[optionr]{function-name}}: }{

     Description of this function.

     Use empty lines for creating vertical space.
   }
   \item{\code{\link[optionr]{function-name}}: }{

     Description of this function.

     Use empty lines for creating vertical space.
   }
}
}
\author{
Who to complain to \email{yourfault@somewhere.net}
}
\references{
\url{http://github.com/whotocomplainto/test.package}
}

