% Generated by roxygen2 (4.0.2): do not edit by hand
\name{reactiveOption}
\alias{reactiveOption}
\title{Create Reactive Option}
\usage{
reactiveOption(expr, where = parent.frame(), ...)
}
\arguments{
\item{expr}{\code{\link{expression}} (quoted or unquoted).}

\item{where}{\code{\link{environment}}.
The parent environment for the reactive expression.
By default, this is the calling environment, the same as when defining
an ordinary non-reactive expression.}

\item{Further}{arguments to be passed along to subsequent functions.
In particular:
\code{\link[reactr]{reactiveExpression}}.}
}
\description{
Creates a reactive option expression. This is a convenience/encapsulation
wrapper around \code{\link[reactr]{reactiveExpression}}.
}
\examples{
\dontrun{

## TODO: add example

}
}
\author{
Janko Thyson \email{janko.thyson@gmail.com}
}
\references{
\url{http://github.com/Rappster/optionr}
}
\seealso{
\code{
    \link[reactr]{reactiveExpression}
}
}

