% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{rmAnywhereOption,character,missing-method}
\alias{rmAnywhereOption,character,missing-method}
\alias{rmAnywhereOption-char-miss-method}
\title{Remove Free Option (char-miss)}
\usage{
\S4method{rmAnywhereOption}{character,missing}(id,
  where = tryCatch(devtools::as.package(".")$package, error = function(cond) {
      stop("Invalid default value for `where`") }), strict = FALSE, ...)
}
\arguments{
\item{id}{\code{\link{character}}.}

\item{where}{\code{\link{missing}}.}

\item{strict}{\code{\link{logical}}.
\code{TRUE}: the following constellations trigger an error:
\itemize{
    \item{\code{id} pointing to a non-existing option}
    \item{empty \code{id}}
}
\code{FALSE}: the stated constellations lead to the return value
being \code{FALSE}.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
See method
   \code{\link[optionr]{rmAnywhereOption-char-char-method}}
}
\description{
See generic: \code{\link[optionr]{rmAnywhereOption}}
}
\examples{
\dontrun{

##------------------------------------------------------------------------------
## Default `where` //
##------------------------------------------------------------------------------

container <- initializeOptionContainer(overwrite = TRUE)
setAnywhereOption(id = "a", value = TRUE)
rmAnywhereOption(id = "a")
exists("a", container, inherits = FALSE)

setAnywhereOption(id = "a/b/c", value = 10, gap = TRUE)
rmAnywhereOption(id = "a/b/c")
exists("c", container$a$b, inherits = FALSE)

##------------------------------------------------------------------------------
## Different `where` // 
##------------------------------------------------------------------------------

where <- "test"
container <- initializeOptionContainer(id = where, overwrite = TRUE)
".test" \%in\% names(options())
## --> option container assigned to R option `.test`

setAnywhereOption(id = "a/b/c", value = 10, where = where, gap = TRUE)
rmAnywhereOption(id = "a/b/c", where = where)
exists("c", container$a$b, inherits = FALSE)

where <- structure(list(id = "test"), class = "OptionContext.Test")
container <- initializeOptionContainer(id = where, overwrite = TRUE)
setAnywhereOption(id = "a/b/c", value = 10, where = where, gap = TRUE)
rmAnywhereOption(id = "a/b/c", where = where)
exists("c", container$a$b, inherits = FALSE)
  
##------------------------------------------------------------------------------
## Strictness //
##------------------------------------------------------------------------------

container <- initializeOptionContainer(overwrite = TRUE)
rmAnywhereOption(id = "a")
try(rmAnywhereOption(id = "a", strict = TRUE))

rmAnywhereOption(id = "a/b/c")
try(rmAnywhereOption(id = "a/b/c", strict = TRUE))

rmAnywhereOption(id = character())
try(rmAnywhereOption(id = character(), strict = TRUE))
    
}
}
\author{
Janko Thyson \email{janko.thyson@gmail.com}
}
\references{
\url{http://github.com/Rappster/optionr}
}
\seealso{
\code{
   \link[optionr]{rmAnywhereOption}
}
}

