% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rmRegistryValue}
\alias{rmRegistryValue}
\title{Remove Registry Value (generic)}
\usage{
rmRegistryValue(id, where = tryCatch(devtools::as.package(".")$package, error
  = function(cond) {     stop("Invalid default value for `where`") }),
  sub_id = character(), strict = c(0, 1, 2), ...)
}
\arguments{
\item{id}{\strong{Signature argument}.
Object containing path-like name/ID information.}

\item{where}{\strong{Signature argument}.
Object containing information about the location of the option container
that is to be used. Typically, this either corresponds to the name/ID
of a package/package project or an instance of a custom class for which
suitable methods in the context of managing options are defined
(see other methods of this package that have signature arguments
\code{id} or \code{where}).}

\item{sub_id}{\code{\link{character}}.
Optional ID for a sub layer. Useful for a hub-like option container
structure.}

\item{strict}{\code{\link{logical}}.
 Controls what happens when \code{id} points to a non-existing registry object:
\itemize{
   \item{0: }{ignore and return \code{FALSE} to signal that the
             assignment process was not successful or \code{fail_value} depending
             on the value of \code{return_status}}
         \item{1: }{ignore and with warning and return \code{FALSE}}
         \item{2: }{ignore and with error}
   }}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\description{
Removes registry value from the respective environment containing registry
information (see \code{\link[optionr]{initializeOptionContainer}}).
}
\section{Path-like identifiers}{


Values for \code{id} are expected to be of structure \code{a/b/c/.../z},
i.e. path-like identifiers using a slash as separator.
The identifier is transformed to \code{a$b$c$...$z} and then in turn to a
valid \emph{get} or \code{assign} expression
(i.e. something similar to \code{getOptionContainer(...)$a$b$c$...$z} and
\code{getOptionContainer(...)$a$b$c$...$z <- value}).
Of course, "atomic" paths (e.g. only \code{a}) are also valid.
}
\examples{
\dontrun{

##------------------------------------------------------------------------------
## Default `where` //
##------------------------------------------------------------------------------

container <- initializeOptionContainer(overwrite = TRUE)
setRegistryValue(id = "a", value = TRUE)
rmRegistryValue(id = "a")
exists("a", container$.registry, inherits = FALSE)

setRegistryValue(id = "a/b/c", value = 10)
rmRegistryValue(id = "a/b/c")
exists("c", container$.registry$a$b, inherits = FALSE)

##------------------------------------------------------------------------------
## Different `where` //
##------------------------------------------------------------------------------
  
where <- "test"
container <- initializeOptionContainer(id = where, overwrite = TRUE)
".test" \%in\% names(options())
setRegistryValue(id = "a/b/c", value = 10, where = where)
rmRegistryValue(id = "a/b/c", where = where)
exists("c", container$.registry$a$b, inherits = FALSE)
  
where <- structure(list(id = "test"), class = "OptionContext.Test")
container <- initializeOptionContainer(id = where, overwrite = TRUE)
".test" \%in\% names(options())
setRegistryValue(id = "a/b/c", value = 10, where = where)
rmRegistryValue(id = "a/b/c", where = where)
exists("c", container$.registry$a$b, inherits = FALSE)

##------------------------------------------------------------------------------
## Strictness //
##------------------------------------------------------------------------------
  
container <- initializeOptionContainer(overwrite = TRUE)
rmRegistryValue(id = "a")
try(rmRegistryValue(id = "a", strict = 2))

rmRegistryValue(id = "a/b/c")
try(rmRegistryValue(id = "a/b/c", strict = 2))

rmRegistryValue(id = character())
try(rmRegistryValue(id = character(), strict = 2))
    
}
}
\author{
Janko Thyson \email{janko.thyson@gmail.com}
}
\references{
\url{http://github.com/Rappster/optionr}
}
\seealso{
\code{
  	\link[optionr]{rmRegistryValue-char-char-method},
    \link[optionr]{setRegistryValue},
    \link[optionr]{getRegistryValue},
    \link[optionr]{existsRegistryValue}
}
}

