% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getOptionContainer}
\alias{getOptionContainer}
\title{Get Option Container (generic)}
\usage{
getOptionContainer(main = tryCatch(devtools::as.package(".")$package, error =
  function(cond) {     stop("Invalid default value for `main`") }),
  hidden = TRUE, ...)
}
\arguments{
\item{main}{\strong{Signature argument}.
Object containing suitable information to control the actual retrieval
process.
In the simplest case, this corresponds
to the name/ID of a package/package project. But it can also be an
instance of a custom class for which a suitable methods are defined.}

\item{hidden}{\code{\link{logical}}.
\code{TRUE}: make sure name/ID information in \code{main}
is preprended with a dot as the container was assigned with \code{hidden = TRUE};
\code{FALSE}: use name/ID information in \code{main} as is.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\description{
Retrieves the \code{environment} that serves as an option
container from the list of R options.
}
\examples{
\dontrun{

##------------------------------------------------------------------------------
## Default `main` //
##------------------------------------------------------------------------------

res <- ensureOptionContainer(overwrite = TRUE)
identical(getOptionContainer(), res)
options(".optionr" = NULL)
  
##------------------------------------------------------------------------------
## Hidden //
##------------------------------------------------------------------------------

res <- ensureOptionContainer(hidden = FALSE, overwrite = TRUE)
identical(getOptionContainer(hidden = FALSE), res)
options("optionr" = NULL)
  
##------------------------------------------------------------------------------
## Interface test //
##------------------------------------------------------------------------------
  
main <- structure(list(id = "test"), class = "OptionContext.Test")
res <- ensureOptionContainer(main, hidden = FALSE, overwrite = TRUE)
identical(getOptionContainer(main, hidden = FALSE), res)
options("test" = NULL)
  
}
}
\author{
Who to complain to \email{yourfault@somewhere.net}
}
\references{
\url{http://github.com/whotocomplainto/test.package}
}
\seealso{
\code{
  	\link[rapp]{getOptionContainer-missing-method}
}
}

