% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getRegistryValue}
\alias{getRegistryValue}
\title{Get Registry Value (generic)}
\usage{
getRegistryValue(id, where = tryCatch(devtools::as.package(".")$package, error
  = function(cond) {     stop("Invalid default value for `where`") }),
  sub_id = character(), default = NULL, strict = c(0, 1, 2), ...)
}
\arguments{
\item{id}{\strong{Signature argument}.
Object containing path-like name/ID information.}

\item{where}{\strong{Signature argument}.
Object containing information about the location of the option container
that is to be used. Typically, this either corresponds to the name/ID
of a package/package project or an instance of a custom class for which
suitable methods in the context of managing options are defined
(see other methods of this package that have signature arguments
\code{id} or \code{where}).}

\item{sub_id}{\code{\link{character}}.
Optional ID for a sub layer. Useful for a hub-like option container
structure.}

\item{default}{\code{\link{ANY}}.
Value to be returned if option does not exist.
See \code{\link[base]{getOption}} and \code{\link[nestr]{getNested}}.}

\item{strict}{\code{\link{logical}}.
 Controls what happens when \code{id} points to a non-existing registry object:
\itemize{
     \item{0: }{ignore and return \code{FALSE} to signal that the
             assignment process was not successful or \code{fail_value} depending
             on the value of \code{return_status}}
         \item{1: }{ignore and with warning and return \code{FALSE}}
         \item{2: }{ignore and with error}
   }}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\description{
Retrieves registry value from the respective environment for registry information
inside an option container (see \code{\link[optionr]{initializeOptionContainer}}).
}
\section{Path-like identifiers}{


Values for \code{id} are expected to be of structure \code{a/b/c/.../z},
i.e. path-like identifiers using a slash as separator.
The identifier is transformed to \code{a$b$c$...$z} and then in turn to a
valid \emph{get} or \code{assign} expression
(i.e. something similar to \code{getOptionContainer(...)$a$b$c$...$z} and
\code{getOptionContainer(...)$a$b$c$...$z <- value}).
Of course, "atomic" paths (e.g. only \code{a}) are also valid.
}
\examples{
\dontrun{

##------------------------------------------------------------------------------
## Default location of option container //
##------------------------------------------------------------------------------

## NOTE
## Obmitting `where` (and `id` in `initializeOptionContainer()`) in the all 
## calls will assign the overall option container to the package/project 
## name (if `hidden = TRUE` it is additionally prepended by a `.`)

initializeOptionContainer(overwrite = TRUE)
setRegistryValue(id = "x_1", value = 10)
getRegistryValue(id = "x_1")

##------------------------------------------------------------------------------
## Explicit location of option container //
##------------------------------------------------------------------------------

initializeOptionContainer(id = "test", overwrite = TRUE)
setRegistryValue(id = "x_1", value = 10, where = "test")
getRegistryValue(id = "x_1", where = "test")

##------------------------------------------------------------------------------
## Sub ID //
##------------------------------------------------------------------------------

initializeOptionContainer(sub_id = "a", overwrite = TRUE)
setRegistryValue(id = "x_1", value = 10, sub_id = "a")
getRegistryValue(id = "x_1", sub_id = "a")

}
}
\author{
Janko Thyson \email{janko.thyson@gmail.com}
}
\references{
\url{http://github.com/Rappster/optionr}
}
\seealso{
\code{
  	\link[optionr]{getRegistryValue-char-char-method},
   \link[optionr]{setRegistryValue},
   \link[optionr]{existsRegistryValue},
   \link[optionr]{rmRegistryValue}
}
}

