% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{initializeMeta,ANY,environment-method}
\alias{initializeMeta,ANY,environment-method}
\alias{initializeMeta-any-env-method}
\title{Initialize Meta Information (any-envir)}
\usage{
\S4method{initializeMeta}{ANY,environment}(id = ".meta", where, ...)
}
\arguments{
\item{id}{\code{\link{ANY}}.}

\item{where}{\code{\link{environment}}.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
\code{\link{environment}}. Environment containing the options.
}
\description{
See generic: \code{\link[optionr]{initializeMeta}}
}
\examples{
\dontrun{

## Default `id` //  
container <- ensureOptionContainer(overwrite = TRUE)  
initializeMeta(where = container)
exists(".meta", container)

## Different `id` //
container <- ensureOptionContainer(overwrite = TRUE)  
initializeMeta(id = "META", where = container)
exists("META", container)

}
}
\author{
Janko Thyson \email{janko.thyson@gmail.com}
}
\references{
\url{http://github.com/Rappster/optionr}
}
\seealso{
\code{
   \link[optionr]{initializeMeta}
}
}

