% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{initializeMeta,character,environment-method}
\alias{initializeMeta,character,environment-method}
\alias{initializeMeta-character-environment-method}
\title{Initialize Meta Information (char-envir)}
\usage{
\S4method{initializeMeta}{character,environment}(main = ".meta", where, ...)
}
\arguments{
\item{main}{\code{\link{character}}.}

\item{where}{\code{\link{environment}}.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
\code{\link{environment}}. Environment containing the options.
}
\description{
See generic: \code{\link[reactr]{initializeMeta}}
}
\examples{
\dontrun{

## TODO: add example

}
}
\author{
Who to complain to \email{yourfault@somewhere.net}
}
\references{
\url{http://github.com/whotocomplainto/test.package}
}
\seealso{
\code{
   \link[reactr]{initializeMeta},
    \link[reactr]{getRegistry}
}
}

