% Generated by roxygen2 (4.0.2): do not edit by hand
\name{initializeOptionContainer}
\alias{initializeOptionContainer}
\title{Initialize Option Container (generic)}
\usage{
initializeOptionContainer(main = tryCatch(devtools::as.package(".")$package,
  error = function(cond) {     stop("Invalid default value for `main`") }),
  components = c("options", ".meta", ".registry"), check = TRUE,
  hidden = TRUE, overwrite = FALSE, ...)
}
\arguments{
\item{main}{\strong{Signature argument}.
Object containing suitable information to control the distinct creation
process.
In the simplest case, this corresponds
to the name/ID of a package/package project. But it can also be an
instance of a class for which methods for
\code{\link[optionr]{initializeOptions}},
\code{\link[optionr]{initializeMeta}} and
\code{\link[optionr]{initializeRegistry}} exist.}

\item{check}{\code{\link{logical}}.
\code{TRUE}: check if an R option with name/ID according to the information
in \code{main} already exists (in which case an error is thrown);
\code{FALSE}: no check for existing R options.
Note that \code{overwrite} will overrule \code{check}.}

\item{hidden}{\code{\link{logical}}.
\code{TRUE}: make sure name/ID information in \code{main}
is preprended with a dot to hide it;
\code{FALSE}: use name/ID information in \code{main} as is.
The former reduces the risk of accidentially overwriting existing R
options and thus is used by default.}

\item{overwrite}{\code{\link{logical}}.
\code{TRUE}: overwrite existing container;
\code{FALSE}: create only if no container exists yet.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\description{
Convenience function to initialize the option container that is then stored
as an R option according to the name/ID provided by \code{id}.
}
\examples{
\dontrun{

##------------------------------------------------------------------------------
## Default `main` //
##------------------------------------------------------------------------------
  
## Make sure initial options are `NULL` //
options("optionr" = NULL)

initializeOptionContainer()
opts <- getOption("optionr")
opts$options  
opts$.meta  
opts$.registry  

## Clean up //
options("optionr" = NULL)

##------------------------------------------------------------------------------
## Explicit character `main` //
##------------------------------------------------------------------------------
  
main <- "test"
options(main = NULL)

initializeOptionContainer(main)
opts <- getOption(main)
opts$options  
opts$.meta  
opts$.registry

## Clean up //
options(main = NULL)
  
##------------------------------------------------------------------------------
## Partial 
##------------------------------------------------------------------------------

main <- "test"
options(main = NULL)
initializeOptionContainer(main, components = "options")
opts <- getOption(main)
opts$options  
opts$.meta
## --> not created
opts$.registry
## --> not created

initializeOptionContainer(main, components = c("options", ".meta"), 
  overwrite = TRUE)
opts <- getOption(main)
opts$options  
opts$.meta
opts$.registry
## --> not created

## Condition handling //
initializeOptionContainer(main, components = c("nonexisting", "options"), 
  overwrite = TRUE)
opts <- getOption(main)
ls(opts, all.names = TRUE)

## Clean up //
options(main = NULL)
  
##------------------------------------------------------------------------------
## As interface //
##------------------------------------------------------------------------------

## Example of how custom classes can be used for creating custom methods //
main <- structure(list(id = "test"), class = "OptionContext.Test")
main

options("test" = NULL)
initializeOptionContainer(main)
## --> calls `initializeOptions()`, `initializeMeta()` and `initializeRegistry()`
## and for each of which methods for signature `main` can be defined

getOption(main$id)$options  
getOption(main$id)$.meta  
getOption(main$id)$.registry  
options("test" = NULL)

}
}
\author{
Who to complain to \email{yourfault@somewhere.net}
}
\references{
\url{http://github.com/whotocomplainto/test.package}
}
\seealso{
\code{
  	\link[optionr]{initializeOptionContainer-NULL-method},
    \link[optionr]{getRegistry}
}
}

