% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{initializeOptions,missing,missing-method}
\alias{initializeOptions,missing,missing-method}
\alias{initializeOptions-missing-missing-method}
\title{Initialize Options (miss-miss)}
\usage{
\S4method{initializeOptions}{missing,missing}(id = "options", where, ...)
}
\arguments{
\item{id}{\code{\link{missing}}.}

\item{where}{\code{\link{missing}}.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
See method
   \code{\link[reactr]{initializeOptions-NULL-method}}.
}
\description{
See generic: \code{\link[reactr]{initializeOptions}}
}
\examples{
\dontrun{

## Default `id` //  
container <- ensureOptionContainer(overwrite = TRUE)  
initializeOptions(where = container)
exists("options", container)

## Different `id` //
container <- ensureOptionContainer(overwrite = TRUE)  
initializeOptions(id = "OPTIONS", where = container)
exists("OPTIONS", container)

}
}
\author{
Who to complain to \email{yourfault@somewhere.net}
}
\references{
\url{http://github.com/whotocomplainto/test.package}
}
\seealso{
\code{
   \link[reactr]{initializeOptions-NULL-method},
    \link[reactr]{getRegistry}
}
}

