% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{initializeRegistry,missing,environment-method}
\alias{initializeRegistry,missing,environment-method}
\alias{initializeRegistry-missing-environment-method}
\title{Initialize Registry (miss-envir)}
\usage{
\S4method{initializeRegistry}{missing,environment}(main = ".registry", where,
  ...)
}
\arguments{
\item{main}{\code{\link{missing}}.}

\item{where}{\code{\link{environment}}.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
See method
   \code{\link[reactr]{initializeRegistry-NULL-method}}.
}
\description{
See generic: \code{\link[reactr]{initializeRegistry}}
}
\examples{
\dontrun{

## TODO: add example

}
}
\author{
Who to complain to \email{yourfault@somewhere.net}
}
\references{
\url{http://github.com/whotocomplainto/test.package}
}
\seealso{
\code{
   \link[reactr]{initializeRegistry-NULL-method},
    \link[reactr]{getRegistry}
}
}

