% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{initializeRegistry,missing,missing-method}
\alias{initializeRegistry,missing,missing-method}
\alias{initializeRegistry-miss-miss-method}
\title{Initialize Registry (miss-miss)}
\usage{
\S4method{initializeRegistry}{missing,missing}(id = ".registry", where, ...)
}
\arguments{
\item{id}{\code{\link{missing}}.}

\item{where}{\code{\link{missing}}.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
See method
   \code{\link[optionr]{initializeRegistry-char-env-method}}.
}
\description{
See generic: \code{\link[optionr]{initializeRegistry}}
}
\examples{
\dontrun{

## Default `id` //  
container <- ensureOptionContainer(overwrite = TRUE)  
initializeRegistry(where = container)
exists(".registry", container)

## Different `id` //
container <- ensureOptionContainer(overwrite = TRUE)  
initializeRegistry(id = "REGISTRY", where = container)
exists("REGISTRY", container)

}
}
\author{
Janko Thyson \email{janko.thyson@gmail.com}
}
\references{
\url{http://github.com/Rappster/optionr}
}
\seealso{
\code{
   \link[optionr]{initializeRegistry}
}
}

