% Generated by roxygen2 (4.0.2): do not edit by hand
\name{initializeRegistry}
\alias{initializeRegistry}
\title{Initialize Registry (generic)}
\usage{
initializeRegistry(main = ".registry", where, ...)
}
\arguments{
\item{main}{\strong{Signature argument}.
Object containing ID information.}

\item{where}{\strong{Signature argument}.
Object containing location information.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\description{
Convenience function to initialize a registry environment inside an
option container.
}
\examples{
\dontrun{

## TODO: add example

}
}
\author{
Who to complain to \email{yourfault@somewhere.net}
}
\references{
\url{http://github.com/whotocomplainto/test.package}
}
\seealso{
\code{
  	\link[reactr]{initializeRegistry-NULL-method},
    \link[reactr]{getRegistry}
}
}

